+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.6] - 2024-10-25
+
+### Changed
+
+- Set default query_type to Intersection for REST API, if not specified in SearchRequestObject
+- Set default similarity_type to Bm25fProximity for REST API, if not specified in CreateIndexRequest
+- Set default tokenizer_type to UnicodeAlphanumeric for REST API, if not specified in CreateIndexRequest
+
 ## [0.7.5] - 2024-10-24
 
 ### Fixed